CHA: modify the way short term correlations are handled

+++ miic.core/src/miic/scripts/calculate_noise_correlations.py
@@ -76,12 +76,12 @@ def paracorr(par):
     logger.addHandler(hdlr) 
     logger.setLevel(logging.DEBUG)
 
+    lle_df = lat_lon_ele_load(par['net']['coordinate_file'])
     comb_list = combine_station_channels(par['net']['stations'],
-                            par['net']['channels'],par['co']['combination_method'])
+                            par['net']['channels'],par['co'],lle_df)
 
     sttimes = datetime_list(par['co']['read_start'], par['co']['read_end'], inc=par['co']['read_inc'])	## loop over 24hrs/whole days
     time_inc = datetime.timedelta(seconds=par['co']['read_len'])
-    lle_df = lat_lon_ele_load(par['net']['coordinate_file'])
     res_dir = par['co']['res_dir']
     station_list = par['net']['stations']
     channel_list = par['net']['channels']
@@ -190,7 +190,7 @@ def paracorr(par):
                     get_valid_traces(sub_st)
                 else:
                     sub_st = st
-                targs['combinations'] = select_available_combinations(sub_st,comb_list,targs)
+                targs.update({'combinations': select_available_combinations(sub_st,comb_list)})
                 if len(targs['combinations']) == 0:
                     continue
                 cst = px.stream_pxcorr(sub_st,targs,comm=comm)
@@ -215,23 +215,19 @@ def paracorr(par):
             
         # if there is a subdivision of read traces
         comm.barrier()
-        if ('subdivision' in par['co']) and (rank == 0):
+        if (('subdivision' in par['co']) and par['co']['subdivision']['recombine_subdivision']) and (rank == 0):
             logger.debug('combining subdivisions')
             # combine traces to matrix
             corr_mat_create_from_traces(pathname, pathname, delete_trace_files=True)
-            if par['co']['subdivision']['recombine_subdivision']:
-                flist = dir_read(pathname,'mat__*.mat')
-                for fl in flist:
-                    try:
-                        mat = mat_to_ndarray(fl)
-                        tr = corr_mat_extract_trace(mat,method='norm_mean')
-                        save_dict_to_matlab_file(fl.replace('mat__','tr__'),tr)
-                        if par['co']['subdivision']['delete_subdivisions']:
-                            os.remove(fl)
-                    except:
-                        pass
-
-
+            flist = dir_read(pathname,'mat__*.mat')
+            for fl in flist:
+                try:
+                    mat = mat_to_ndarray(fl)
+                    tr = corr_mat_extract_trace(mat,method='norm_mean')
+                    save_dict_to_matlab_file(fl.replace('mat__','tr__'),tr)
+                    os.remove(fl)
+                except:
+                    pass
 
     program_end = UTCDateTime()
 
@@ -268,70 +264,41 @@ def merge_corr_results(par):
     print '\nrank %d of %d'  % (rank,psize)
     logger.debug('Rank %d of %d Beginning execution.'  % (rank,psize))
 
+    lle_df = lat_lon_ele_load(par['net']['coordinate_file'])
     comb_list = combine_station_channels(par['net']['stations'],
-                            par['net']['channels'],par['co']['combination_method'])
+                            par['net']['channels'],par['co'],lle_df)
 
-    sttimes = datetime_list(par['co']['read_start'], par['co']['read_end'], inc=par['co']['read_inc'])
+    if ('subdivision' in par['co']) and not par['co']['subdivision']['recombine_subdivision']:
+        sttimes = datetime_list(par['co']['read_start'], par['co']['read_end'], 
+                                inc=par['co']['subdivision']['corr_inc'])
+    else:
+        sttimes = datetime_list(par['co']['read_start'], par['co']['read_end'], inc=par['co']['read_inc'])
+
 
     # mapping of combinations to processes
     pmap = (np.arange(len(comb_list[0]))*psize)/len(comb_list[0])
     # indecies for stations to be worked on by each process
     comb_ind = np.where(pmap == rank)[0]
-    #pdb.set_trace()
 
-    # combine matrices of subdivisions
-    if ('subdivision' in par['co']) and not par['co']['subdivision']['recombine_subdivision']:
-        for cind in comb_ind:
-            print "\n>>> Working on combination %s-%s at %s:" % \
-                        (comb_list[0][cind],comb_list[1][cind],UTCDateTime())
-            logger.debug("\n>>> Working on combination %s-%s at %s:" % \
-                        (comb_list[0][cind],comb_list[1][cind],UTCDateTime()))
-            ID0 = comb_list[0][cind].split('.')
-            ID1 = comb_list[1][cind].split('.')
-            comb_str = 'mat__%s%s.%s%s.*.%s%s.mat' % (ID0[0],ID1[0],ID0[1],ID1[1],ID0[3],ID1[3])
-            matl = []
-            year_list = dir_read(par['co']['res_dir'],'????')
-            year_list = sorted(year_list)
-            for year in year_list:
-                day_list = dir_read(year,'???')
-                day_list = sorted(day_list)
-                for day in day_list:
+    for cind in comb_ind:
+        logger.debug("\n>>> Working on combination %s-%s at %s:" % \
+                    (comb_list[0][cind],comb_list[1][cind],UTCDateTime()))
+        ID0 = comb_list[0][cind].split('.')
+        ID1 = comb_list[1][cind].split('.')
+        comb_str = '*_%s%s.%s%s.*.%s%s.mat' % (ID0[0],ID1[0],ID0[1],ID1[1],ID0[3],ID1[3])
+        trl = []
+        year_list = dir_read(par['co']['res_dir'],'????')
+        year_list = sorted(year_list)
+        for year in year_list:
+            day_list = dir_read(year,'???')
+            day_list = sorted(day_list)
+            for day in day_list:
+                flist = glob.glob(os.path.join(day,comb_str))
+                flist = sorted(flist)
+                for fname in flist:
                     try:
-                        fname = glob.glob(os.path.join(day,comb_str))[0]
                         # the following line avoids including traces after a change in the location code
-                        comb_str = os.path.split(fname)[1]
-                        mat = mat_to_ndarray(fname)
-                        matl.append(mat)
-                    except:
-                        logger.warning("Problem with fname: %s" % 
-                                        (comb_list[0][cind], comb_list[1][cind], sys.exc_info()[0]))
-            try:
-                mat = corr_mat_merge(matl)
-                tr = corr_mat_extract_trace(mat,method='norm_mean')
-                save_dict_to_matlab_file(os.path.join(par['co']['res_dir'],comb_str.replace('mat__','ctr__')),tr)
-                rmat = corr_mat_resample(mat,sttimes)
-                save_dict_to_matlab_file(os.path.join(par['co']['res_dir'],comb_str),rmat)
-            except:
-                logger.warning("Problem with combination %s-%s: %s" % 
-                                (comb_list[0][cind], comb_list[1][cind], sys.exc_info()[0]))
-
-    # No subdivisions or subdivisions already recombined
-    else:
-        for cind in comb_ind:
-            ID0 = comb_list[0][cind].split('.')
-            ID1 = comb_list[1][cind].split('.')
-            comb_str = 'tr__%s%s.%s%s.*.%s%s.mat' % (ID0[0],ID1[0],ID0[1],ID1[1],ID0[3],ID1[3])
-            trl = []
-            year_list = dir_read(par['co']['res_dir'],'????')
-            year_list = sorted(year_list)
-            for year in year_list:
-                day_list = dir_read(year,'???')
-                day_list = sorted(day_list)
-                for day in day_list:
-                    try:
-                        fname = glob.glob(os.path.join(day,comb_str))[0]
-                        # the following line avoids including traces after a change in the location code
-                        comb_str = os.path.split(fname)[1]
+                        ID_str = os.path.split(fname)[1].split('_')[-1]
                         tr = mat_to_ndarray(fname)
                         if tr['stats']['sampling_rate'] != float(par['co']['sampling_rate'])/par['co']['decimation']:
                             if np.abs(np.log10(tr['stats']['sampling_rate']/float(par['co']['sampling_rate'])/par['co']['decimation']))<1e-6:
@@ -342,16 +309,17 @@ def merge_corr_results(par):
                         trl.append(tr)
                     except:
                         pass
-            try:
-                st = corr_trace_to_obspy(trl)
-                mat = corr_mat_from_corr_stream(st)
-                tr = corr_mat_extract_trace(mat,method='norm_mean')
-                save_dict_to_matlab_file(os.path.join(par['co']['res_dir'],comb_str.replace('tr__','ctr__')),tr)
-                rmat = corr_mat_resample(mat,sttimes)
-                save_dict_to_matlab_file(os.path.join(par['co']['res_dir'],comb_str.replace('tr__','mat__')),rmat)
-            except:
-                logger.warning("Problem with combination %s-%s: %s" % 
-                                (comb_list[0][cind], comb_list[1][cind], sys.exc_info()[0]))
+        try:
+            st = corr_trace_to_obspy(trl)
+            mat = corr_mat_from_corr_stream(st)
+            tr = corr_mat_extract_trace(mat,method='norm_mean')
+            save_dict_to_matlab_file(os.path.join(par['co']['res_dir'],'ctr__'+ID_str),tr)
+            rmat = corr_mat_resample(mat,sttimes)
+            rmat = mat
+            save_dict_to_matlab_file(os.path.join(par['co']['res_dir'],'mat__'+ID_str),rmat)
+        except:
+            logger.warning("Problem with combination %s-%s: %s" % 
+                            (comb_list[0][cind], comb_list[1][cind], sys.exc_info()[0]))
 
 
     program_end = UTCDateTime()
@@ -371,7 +339,6 @@ if __name__=="__main__":
     
     # create output directory
     create_path(par['co']['res_dir'])
-
     # do correlations
     paracorr(par)
     # combine correlation results
